src/requirements.txt
flask
flask-socketio
websocket-client
requests
python-dotenv==0.19.1